refactor: 각 클래스 도메인에 맞게 재배치
