+++ src/main/java/com/yunsseong/ai_ex_server/auth/dto/SignUpRequest.java
@@ -1,4 +1,4 @@
-package com.yunsseong.ai_ex_server.member.application.dto;
+package com.yunsseong.ai_ex_server.auth.dto;
 
 public record SignUpRequest(String nickname, String email, String password1, String password2) {
 }
