+++ src/main/java/com/yunsseong/ai_ex_server/common/exception/CommonStatusCode.java
@@ -1,4 +1,4 @@
-package com.yunsseong.ai_ex_server.common.exception.error_code;
+package com.yunsseong.ai_ex_server.common.exception;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
