+++ src/main/java/com/yunsseong/ai_ex_server/common/exception/SystemException.java
@@ -1,6 +1,5 @@
 package com.yunsseong.ai_ex_server.common.exception;
 
-import com.yunsseong.ai_ex_server.common.exception.error_code.StatusConst;
 import lombok.Getter;
 
 @Getter
